Split settings into separate components
